Extract `expect(_:toCompleteFeedLoadWith:)` to a shared scope to remove duplication
