+++ FeedFetchApp/FeedFetchAppTests/FeedLoaderCacheDecoratorTests.swift
@@ -26,7 +26,7 @@ final class FeedLoaderCacheDecoratorTests: XCTestCase {
         let loader = FeedLoaderStub(result: .success(feed))
         let sut = FeedLoaderCacheDecorator(decoratee: loader)
                         
-        expect(sut, toCompleteWith: .success(feed))
+        expect(sut, toCompleteFeedLoadWith: .success(feed))
     }
     
     func test_load_deliversErrorOnLoaderFailure() {
@@ -34,32 +34,8 @@ final class FeedLoaderCacheDecoratorTests: XCTestCase {
         let loader = FeedLoaderStub(result: .failure(loaderError))
         let sut = FeedLoaderCacheDecorator(decoratee: loader)
                         
-        expect(sut, toCompleteWith: .failure(loaderError))
+        expect(sut, toCompleteFeedLoadWith: .failure(loaderError))
     }
     
     
-}
-
-// MARK:- Helpers
-
-private extension FeedLoaderCacheDecoratorTests {
-        
-    func expect(_ sut: FeedLoaderCacheDecorator, toCompleteWith expected: FeedLoader.Result, file: StaticString = #file, line: UInt = #line) {
-        let exp = expectation(description: "Waiting for load to complete")
-        
-        sut.load { result in
-            switch (result, expected) {
-            case let (.success(resultFeed), .success(expectedFeed)):
-                XCTAssertEqual(resultFeed, expectedFeed, "Expected load to complete with feed \(expectedFeed) but got \(resultFeed)", file: file, line: line)
-            case let (.failure(resultError as NSError), .failure(expectedError as NSError)):
-                XCTAssertEqual(resultError, expectedError, "Expected load to fail with error \(expectedError) but got \(resultError)", file: file, line: line)
-            default:
-                XCTFail("Expected load to complete with result \(expected) but got \(result)")
-            }
-                                    
-            exp.fulfill()
-        }
-        
-        wait(for: [exp], timeout: 1.0)
-    }
 }
