+++ FeedFetchApp/FeedFetchAppTests/FeedLoaderWithFallbackCompositeTests.swift
@@ -48,23 +48,4 @@ private extension FeedLoaderWithFallbackCompositeTests{
         return sut
     }
     
-    func expect(_ sut: FeedLoaderWithFallbackComposite, toCompleteFeedLoadWith expected: FeedLoader.Result,file: StaticString = #file, line: UInt = #line) {
-        let exp = expectation(description: "Waiting for feed load to complete")
-        
-        sut.load { result in
-            switch (result, expected) {
-            case let (.failure(error as NSError), .failure(expectedError as NSError)):
-                XCTAssertEqual(expectedError, error, "Expected feed load to fail with error \(expectedError) but got \(error)")
-            case let (.success(resultFeed), .success(expectedFeed)):
-                XCTAssertEqual(expectedFeed, resultFeed, "Expected feed load to complete with feed \(expectedFeed) but got \(resultFeed)")
-            default:
-                XCTFail("Expecte load feed to complete with \(expected) but got \(result)")
-            }
-                        
-            exp.fulfill()
-        }
-        
-        wait(for: [exp], timeout: 1.0)
-    }
-    
 }
